+++ src/diagnostic.rs
@@ -98,6 +98,26 @@ pub enum Applicability {
     Unknown,
 }
 
+/// The diagnostic level
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum DiagnosticLevel {
+    /// Internal compiler error
+    #[serde(rename = "error: internal compiler error")]
+    Ice,
+    /// Error
+    Error,
+    /// Warning
+    Warning,
+    /// Note
+    Note,
+    /// Help
+    Help,
+    #[doc(hidden)]
+    #[serde(other)]
+    Unknown,
+}
+
 /// A diagnostic message generated by rustc
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Diagnostic {
@@ -106,7 +126,7 @@ pub struct Diagnostic {
     /// The associated error code for this diagnostic
     pub code: Option<DiagnosticCode>,
     /// "error: internal compiler error", "error", "warning", "note", "help"
-    pub level: String,
+    pub level: DiagnosticLevel,
     /// A list of source code spans this diagnostic is associated with.
     pub spans: Vec<DiagnosticSpan>,
     /// Associated diagnostic messages.
@@ -116,4 +136,14 @@ pub struct Diagnostic {
     #[doc(hidden)]
     #[serde(skip)]
     __do_not_match_exhaustively: (),
+}
+
+impl ToString for Diagnostic {
+    fn to_string(&self) -> String {
+        if let Some(ref rendered) = self.rendered {
+            rendered.clone()
+        } else {
+            "cargo didn't render this message".to_string()
+        }
+    }
 }
